client/requirements.txt
tqdm>=4.62.3
weaviate-client>=3.3.0

